requirements.txt
Django==1.5.1
MySQL-python==1.2.4
South==0.8.1
argparse==1.2.1
django-tastypie==0.9.15
#ipython==0.13.2
mimeparse==0.1.3
#psycopg2==2.5.1
#python-dateutil==2.1
#six==1.3.0
#wsgiref==0.1.2
